move to scroll state to view from scroll view
